Fix time range copying and add byte range copying
git-svn-id: https://svn.atnf.csiro.au/difx@2262 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ applications/mk5daemon/branches/difx-1.5/mk5dir/mark5dir.cpp
@@ -910,7 +910,7 @@ int getByteRange(const struct Mark5Scan *scan, long long *byteStart, long long *
 	}
 	else
 	{
-		*byteStart = (int)(scan->start + R*(mjdStart - scanStart));
+		*byteStart = (long long)(scan->start + R*(mjdStart - scanStart));
 	}
 
 	if(mjdStop >= scanStop)
@@ -919,7 +919,7 @@ int getByteRange(const struct Mark5Scan *scan, long long *byteStart, long long *
 	}
 	else
 	{
-		*byteStop = (int)(scan->start + R*(mjdStop - scanStart));
+		*byteStop = (long long)(scan->start + R*(mjdStop - scanStart));
 	}
 
 	/* make sure read is aligned with data frames */
